w2c/requirements.txt
pandas>=2.0.0
pyarrow>=12.0.0
PyYAML>=6.0
tqdm>=4.0
Pillow>=9.0

# Optional (depending on which stages you run):
# torch
# torchvision
# vllm
# easyocr
# spacy
# nltk
